+++ OpenGL/src/util/Spline.hpp
@@ -9,41 +9,37 @@ static const float smoothing = 0.2f;
 */
 class Spline {
 public:
-    Spline(std::pair<float, Bezier> bezier);
-    Spline(std::vector<std::pair<float, Bezier> > beziers);
-    Spline(std::vector<std::pair<float, glm::vec3> > keyframes);
-    Spline(std::vector<std::pair<float, glm::vec3> > keyframes, glm::vec3 initialVector, glm::vec3 finalVector);
+    Spline(std::pair<uint32_t, Bezier> bezier);
+    Spline(std::vector<std::pair<uint32_t, Bezier> > beziers);
+    Spline(std::vector<std::pair<uint32_t, glm::vec3> > keyframes);
+    Spline(std::vector<std::pair<uint32_t, glm::vec3> > keyframes, glm::vec3 initialVector, glm::vec3 finalVector);
 
-    glm::vec3 compute(float &dt);
-
-    //Deprecated
-    // void add(std::vector<std::pair<float, Bezier>> beziers);
-    // void add(std::pair<float, Bezier> bezier);
+    glm::vec3 compute(uint32_t &dt);
 
     void reset();
 
-    float getTotalTime(){return totalTime;}
-    float getSplineTime(){return splineTime;}
+    uint32_t getTotalTime(){return totalTime;}
+    uint32_t getSplineTime(){return splineTime;}
     
     glm::vec3 getCurrentPoint();
     glm::vec3 getCurrentDerivative();
     float getCurrentSpeed();
     
-    glm::vec3 getPoint(float t);
-    glm::vec3 getDerivative(float t);
-    float getSpeed(float t);
+    glm::vec3 getPoint(uint32_t t);
+    glm::vec3 getDerivative(uint32_t t);
+    float getSpeed(uint32_t t);
 
     float getLength(int bezierPrecision);
 
 private:
     //Duration, Bezier
-    std::vector<std::pair<float, Bezier> > beziers;
+    std::vector<std::pair<uint32_t, Bezier> > beziers;
     
     size_t currentBezier;
-    float currentBezierTime;
+    uint32_t currentBezierTime;
     
-    float splineTime;
-    float totalTime;
+    uint32_t splineTime;
+    uint32_t totalTime;
 
     bool finished;
 };
